refactor: removing revalidateTag on server actions createEntry and deleteEntry.

+++ src/actions/entry/deleteEntry.ts
@@ -1,6 +1,6 @@
 'use server'
 import prisma from '@/lib/prisma'
-import { revalidateTag } from 'next/cache'
+import { revalidatePath } from 'next/cache'
 import { z } from 'zod'
 
 export const deleteEntry = async (data: FormData) => {
@@ -12,11 +12,30 @@ export const deleteEntry = async (data: FormData) => {
     id: data.get('id'),
   })
 
-  await prisma.entry.delete({
+  const entry = await prisma.entry.findUnique({
     where: {
       id,
     },
   })
 
-  revalidateTag('tasks')
+  if (entry) {
+    await prisma.entry.delete({
+      where: {
+        id,
+      },
+    })
+
+    await prisma.task.update({
+      where: {
+        id: entry.taskId,
+      },
+      data: {
+        progress: {
+          decrement: entry.increment,
+        },
+      },
+    })
+  }
+
+  revalidatePath(`/task/${entry?.taskId}`, 'page')
 }
